+++ src/main/java/br/com/zalf/prolog/webservice/integracao/avacorpavilan/AvaCorpAvilan.java
@@ -2,6 +2,7 @@ package br.com.zalf.prolog.webservice.integracao.avacorpavilan;
 
 import br.com.zalf.prolog.webservice.colaborador.model.Colaborador;
 import br.com.zalf.prolog.webservice.frota.checklist.model.Checklist;
+import br.com.zalf.prolog.webservice.frota.checklist.model.FarolChecklist;
 import br.com.zalf.prolog.webservice.frota.checklist.model.NovoChecklistHolder;
 import br.com.zalf.prolog.webservice.frota.checklist.modelo.ModeloChecklist;
 import br.com.zalf.prolog.webservice.frota.pneu.afericao.model.Afericao;
@@ -13,12 +14,14 @@ import br.com.zalf.prolog.webservice.frota.veiculo.model.Veiculo;
 import br.com.zalf.prolog.webservice.frota.veiculo.model.diagrama.DiagramaVeiculo;
 import br.com.zalf.prolog.webservice.integracao.IntegradorProLog;
 import br.com.zalf.prolog.webservice.integracao.avacorpavilan.cadastro.ArrayOfVeiculo;
+import br.com.zalf.prolog.webservice.integracao.avacorpavilan.checklist.ArrayOfFarolDia;
 import br.com.zalf.prolog.webservice.integracao.avacorpavilan.checklist.ArrayOfVeiculoQuestao;
 import br.com.zalf.prolog.webservice.integracao.avacorpavilan.requester.AvaCorpAvilanRequester;
 import br.com.zalf.prolog.webservice.integracao.sistema.Sistema;
 import com.sun.istack.internal.NotNull;
 import com.sun.istack.internal.Nullable;
 
+import java.util.Date;
 import java.util.List;
 import java.util.Map;
 
@@ -51,11 +54,46 @@ public final class AvaCorpAvilan extends Sistema {
     }
 
     @Override
-    public boolean insertChecklist(@NotNull Checklist checklist) throws Exception {
+    public NovoChecklistHolder getNovoChecklistHolder(@NotNull Long codUnidade,
+                                                      @NotNull Long codModelo,
+                                                      @NotNull String placaVeiculo,
+                                                      char tipoChecklist) throws Exception {
+        final ArrayOfVeiculoQuestao questoesVeiculo = requester.getQuestoesVeiculo(
+                Math.toIntExact(codModelo),
+                placaVeiculo,
+                AvacorpAvilanTipoChecklist.fromTipoProLog(tipoChecklist),
+                cpf(),
+                dataNascimento());
+        return AvaCorpAvilanConverter.convert(questoesVeiculo, placaVeiculo);
+    }
+
+    @Override
+    public Long insertChecklist(@NotNull Checklist checklist) throws Exception {
         return requester.insertChecklist(
                 AvaCorpAvilanConverter.convert(checklist, cpf(), dataNascimento()),
                 cpf(),
                 dataNascimento());
+    }
+
+    @Override
+    public Checklist getByCod(Long codChecklist) throws Exception {
+        throw new UnsupportedOperationException();
+    }
+
+    @Override
+    public FarolChecklist getFarolChecklist(@NotNull final Long codUnidade,
+                                            @NotNull final Date dataInicial,
+                                            @NotNull final Date dataFinal,
+                                            final boolean itensCriticosRetroativos) throws Exception {
+        final String codUnidadeAvilan = getIntegradorProLog().getCodUnidadeClienteByCodUnidadeProLog(codUnidade);
+        final ArrayOfFarolDia farolChecklist = requester.getFarolChecklist(
+                Integer.parseInt(codUnidadeAvilan),
+                AvaCorpAvilanUtils.createDatePattern(dataInicial),
+                AvaCorpAvilanUtils.createDatePattern(dataFinal),
+                itensCriticosRetroativos,
+                cpf(),
+                dataNascimento());
+        return AvaCorpAvilanConverter.convert(farolChecklist);
     }
 
     @Override
@@ -71,6 +109,9 @@ public final class AvaCorpAvilan extends Sistema {
         final List<Pneu> pneus = AvaCorpAvilanConverter.convert(requester.getPneusVeiculo(placaVeiculo, cpf(), dataNascimento()));
         final Restricao restricao = getIntegradorProLog().getRestricaoByCodUnidade(codUnidade());
         final DiagramaVeiculo diagramaVeiculo = getIntegradorProLog().getDiagramaVeiculoByPlaca(placaVeiculo);
+        if (diagramaVeiculo == null) {
+            throw new IllegalStateException("Diagrama não encontrado para a placa: " + placaVeiculo);
+        }
         veiculo.setDiagrama(diagramaVeiculo);
         veiculo.setListPneus(pneus);
 
@@ -89,24 +130,12 @@ public final class AvaCorpAvilan extends Sistema {
         return requester.insertAfericao(AvaCorpAvilanConverter.convert(afericao), cpf(), dataNascimento());
     }
 
-    @Override
-    public NovoChecklistHolder getNovoChecklistHolder(@NotNull Long codUnidade,
-                                                      @NotNull Long codModelo,
-                                                      @NotNull String placaVeiculo) throws Exception {
-        final ArrayOfVeiculoQuestao questoesVeiculo = requester.getQuestoesVeiculo(
-                Math.toIntExact(codModelo),
-                placaVeiculo,
-                cpf(),
-                dataNascimento());
-        return AvaCorpAvilanConverter.convert(questoesVeiculo, placaVeiculo);
-    }
-
     private String cpf() throws Exception {
         if (colaborador == null) {
             colaborador = getIntegradorProLog().getColaboradorByToken(getUserToken());
         }
 
-        // Preenche com 0 a esquerda caso CPF tenha menos do que 11 caracteres
+        // Preenche com 0 a esquerda caso CPF tenha menos do que 11 caracteres.
         return String.format("%011d",  colaborador.getCpf());
     }