+++ src/main/java/br/com/zalf/prolog/webservice/integracao/avacorpavilan/checklist/RespostasAvaliacao.java
@@ -148,5 +148,4 @@ public class RespostasAvaliacao {
     public void setRespostas(ArrayOfRespostaAval value) {
         this.respostas = value;
     }
-
 }