+++ src/main/java/br/com/zalf/prolog/webservice/integracao/operacoes/OperacoesIntegradasChecklist.java
@@ -1,10 +1,12 @@
 package br.com.zalf.prolog.webservice.integracao.operacoes;
 
 import br.com.zalf.prolog.webservice.frota.checklist.model.Checklist;
+import br.com.zalf.prolog.webservice.frota.checklist.model.FarolChecklist;
 import br.com.zalf.prolog.webservice.frota.checklist.model.NovoChecklistHolder;
 import br.com.zalf.prolog.webservice.frota.checklist.modelo.ModeloChecklist;
 import com.sun.istack.internal.NotNull;
 
+import java.util.Date;
 import java.util.List;
 import java.util.Map;
 
@@ -20,7 +22,17 @@ interface OperacoesIntegradasChecklist {
     NovoChecklistHolder getNovoChecklistHolder(
             @NotNull final Long codUnidade,
             @NotNull final Long codModelo,
-            @NotNull final String placaVeiculo) throws Exception;
+            @NotNull final String placaVeiculo,
+            final char tipoChecklist) throws Exception;
 
-    boolean insertChecklist(@NotNull final Checklist checklist) throws Exception;
+    Long insertChecklist(@NotNull final Checklist checklist) throws Exception;
+
+    @NotNull
+    Checklist getByCod(@NotNull final Long codChecklist) throws Exception;
+
+    @NotNull
+    FarolChecklist getFarolChecklist(@NotNull final Long codUnidade,
+                                     @NotNull final Date dataInicial,
+                                     @NotNull final Date dataFinal,
+                                     final boolean itensCriticosRetroativos) throws Exception;
 }